+++ src/template.ts
@@ -1,8 +1,8 @@
 
 // Models
 import { InputArguments } from "./models/InputArguments";
-import { TypeInfo, FieldInfo, PropertyInfo, EventInfo, MethodInfo, QuickTypeInfo } from "./models/SharpChecker";
-import { TemplateApiItems, TemplateApiUris, MemberList } from "./models/TemplateApi";
+import { TypeInfo, FieldInfo, PropertyInfo, EventInfo, MethodInfo, QuickTypeInfo, ParameterInfo, GenericParametersInfo } from "./models/SharpChecker";
+import { TemplateApiItems, TemplateApiUris, MemberList, NameDescription, ParameterNameDescription, GenericParameterNameDescription } from "./models/TemplateApi";
 import { SidebarView } from "./models/TemplateApi";
 import { XmlFormat } from "./models/XmlFormat";
 // External functionality
@@ -17,6 +17,9 @@ import prettier = require("prettier");
 
 // Variables
 let generatedTypeJson : TypeInfo;
+
+// TODO: Add a compileHeader function to compile the header.
+// TODO: Add a compileFooter function to compile the footer.
 
 /**Compiles the sidebar for it's own separate file.
  * @param args {InputArguments} - The input arguments to look into for the uris.
@@ -102,8 +105,8 @@ export function compileType(filename : string, typePath : string) : string {
 	// Variables
 	const args : InputArguments = getArguments();
 	const xmlFormat = getApiDoc(`T:${ typePath }`, getDependencies());
-	const xmlApi : TemplateApiItems = getApiItems(xmlFormat);
 	const details = generatedTypeJson;
+	const xmlApi : TemplateApiItems = getApiItems(xmlFormat, details);
 	const uris = {
 		constructors: args.templateUris.constructors,
 		fields: args.templateUris.fields,
@@ -141,7 +144,7 @@ export function compileField(filename : string, details : FieldInfo) {
 	// Variables
 	const typePath = getTypePath(details.implementedType, details.name);
 	const xmlFormat = getApiDoc(`F:${ typePath }`, getDependencies());
-	const xmlApi : TemplateApiItems = getApiItems(xmlFormat);
+	const xmlApi : TemplateApiItems = getApiItems(xmlFormat, details);
 	
 	return ejs.render(
 		readFile(filename).replace(/\s+\n/gm, "\n").replace(/\t/gm, "  ").trim(),
@@ -167,7 +170,7 @@ export function compilePropety(filename : string, details : PropertyInfo) {
 	// Variables
 	const typePath = getPropertyTypePath(details);
 	const xmlFormat = getApiDoc(`P:${ typePath }`, getDependencies());
-	const xmlApi : TemplateApiItems = getApiItems(xmlFormat);
+	const xmlApi : TemplateApiItems = getApiItems(xmlFormat, details);
 	
 	return ejs.render(
 		readFile(filename).replace(/\s+\n/gm, "\n").replace(/\t/gm, "  ").trim(),
@@ -193,7 +196,7 @@ export function compileEvent(filename : string, details : EventInfo) {
 	// Variables
 	const typePath = getTypePath(details.implementedType, details.name);
 	const xmlFormat = getApiDoc(`E:${ typePath }`, getDependencies());
-	const xmlApi : TemplateApiItems = getApiItems(xmlFormat);
+	const xmlApi : TemplateApiItems = getApiItems(xmlFormat, details);
 	
 	return ejs.render(
 		readFile(filename).replace(/\s+\n/gm, "\n").replace(/\t/gm, "  ").trim(),
@@ -219,7 +222,7 @@ export function compileMethod(filename : string, details : MethodInfo) {
 	// Variables
 	const typePath = getMethodTypePath(details);
 	const xmlFormat = getApiDoc(`M:${ typePath }`, getDependencies());
-	const xmlApi : TemplateApiItems = getApiItems(xmlFormat);
+	const xmlApi : TemplateApiItems = getApiItems(xmlFormat, details);
 		
 	return ejs.render(
 		readFile(filename).replace(/\s+\n/gm, "\n").replace(/\t/gm, "  ").trim(),
@@ -383,15 +386,21 @@ function getTypePath(typeInfo : QuickTypeInfo, name : string) : string {
 /**Gets the api documentation items to be used by the template.
  * @param format {XmlFormat|undefined} - The format used to get the documentation from.
  * @return Returns the api documentation items.*/
-function getApiItems(format : (XmlFormat | undefined)) : TemplateApiItems {
+function getApiItems(format : (XmlFormat | undefined), details : (TypeInfo | MethodInfo | PropertyInfo | EventInfo | FieldInfo)) : TemplateApiItems {
 	if(format == undefined) { format = new XmlFormat(); }
-	// TODO: Add details parameter.
+	
 	return {
 		summary: format.summary,
-		// TODO: Update this to reflect the details iff the format doesn't exist
 		returns: {
-			exists: doesItemExist(format.returns),
-			value: format.returns
+			exists: doesItemExist(
+				format.returns ||
+				((details as MethodInfo).returnType && (details as MethodInfo).returnType.name != "void" ?
+					(details as MethodInfo).returnType.name :
+					""
+				)
+			),
+			value: format.returns,
+			details: (details as MethodInfo).returnType
 		},
 		remarks: {
 			exists: doesItemExist(format.remarks),
@@ -401,22 +410,106 @@ function getApiItems(format : (XmlFormat | undefined)) : TemplateApiItems {
 			exists: doesItemExist(format.example),
 			value: format.example
 		},
-		// TODO: Update this to reflect the details iff the format doesn't exist
-		// TODO: Update the value to fill up by the details
 		parameters: {
-			exists: doesArrayItemExist(format.parameters),
-			value: format.parameters
+			exists: doesArrayItemExist(format.parameters || (details as (MethodInfo | PropertyInfo)).parameters),
+			value: getParameterDetails(format.parameters, details)
 		},
 		exceptions: {
 			exists: doesArrayItemExist(format.exceptions),
 			value: format.exceptions
 		},
-		// TODO: Update this to reflect the details iff the format doesn't exist
 		typeParameters: {
-			exists: doesArrayItemExist(format.typeParameters),
-			value: format.typeParameters
+			exists: doesArrayItemExist(
+				format.typeParameters ||
+				(details as MethodInfo).genericParameters ||
+				((details as TypeInfo).typeInfo ?
+					(details as TypeInfo).typeInfo.genericParameters :
+					undefined
+				)
+			),
+			value: getGenericParameterDetails(format.typeParameters, details)
 		}
 	};
+}
+
+/**Gets all the details of the method/property's parameters with the added description found within the xml. This is
+ * so that even if there is no XML documentation to the parameters, they will still appear in the template.
+ * @param format {NameDescription[]} - The list of names and descriptions of the parameters that were gathered.
+ * @param details {MethodInfo | PropertyInfo | any} - The details to look into.
+ * @returns Returns a true list of parameter details with descriptions, whether or not documented in the XML.*/
+function getParameterDetails(format : NameDescription[], details : (MethodInfo | PropertyInfo | any)) : ParameterNameDescription[] {
+	if(!details || !(details as (MethodInfo | PropertyInfo)).parameters) { return []; }
+	
+	// Variables
+	let parameters : ParameterNameDescription[] = [];
+	let temp : ParameterInfo[] = (details as (MethodInfo | PropertyInfo)).parameters;
+	
+	for(let a = 0; a < temp.length; a++) {
+		// Variables
+		let index = -1;
+		let parameter : ParameterNameDescription = {
+			name: temp[a].name,
+			description: "<p>No description.</p>",
+			details: temp[a]
+		};
+		
+		for(let b = 0; b < format.length; b++) {
+			if(temp[a].name == format[b].name) {
+				index = b;
+				break;
+			}
+		}
+		
+		if(index != -1) { parameter.description = format[index].description; }
+		parameters.push(parameter);
+	}
+	
+	return parameters;
+}
+
+/**Gets all the details of the type/method's generic parameters with the added description found within the xml. This
+ * is so that even if there is no XML documentation to the generic parameters, they will still appear in the template.
+ * @param format {NameDescription[]} - The list of names and description of the generic parameters that were gathered.
+ * @param details {TypeInfo | MethodInfo | any} - The details to look into.
+ * @returns Returns a true list of generic parameter details with descriptions, whether or not documented in the XML.*/
+function getGenericParameterDetails(format : NameDescription[], details : (TypeInfo | MethodInfo | any)) : GenericParameterNameDescription[] {
+	if(
+		!details ||
+		!(details as MethodInfo).genericParameters ||
+		!(details as TypeInfo).typeInfo ||
+		!(details as TypeInfo).typeInfo.genericParameters
+	) {
+		return [];
+	}
+	
+	// Variables
+	let typeParameters : GenericParameterNameDescription[] = [];
+	let temp : GenericParametersInfo[] = ((details as MethodInfo).genericParameters ?
+		(details as MethodInfo).genericParameters :
+		(details as TypeInfo).typeInfo.genericParameters
+	);
+	
+	for(let a = 0; a < temp.length; a++) {
+		// Variables
+		let index = -1;
+		let generic : GenericParameterNameDescription = {
+			name: temp[a].name,
+			description: "<p>No description.</p>",
+			details: temp[a]
+		};
+		
+		for(let b = 0; b < format.length; b++) {
+			if(temp[a].name == format[b].name) {
+				index = b;
+				break;
+			}
+		}
+		
+		if(index != -1) { generic.description = format[index].description; }
+		typeParameters.push(generic);
+	}
+	
+	return typeParameters;
 }
 
 /**Finds if the given string is non-empty and exists.
@@ -427,9 +520,7 @@ function doesItemExist(str : string) : boolean { return (str != null && str != u
 /**Finds if the array is non-empty and exists.
  * @param list {any[]} - The list in question.
  * @returns Returns true if the list is non-empty and exists.*/
-function doesArrayItemExist(list : any[]) : boolean {
-	return (list != null && list != undefined && list.length > 0);
-}
+function doesArrayItemExist(list : any[]) : boolean { return (list != null && list != undefined && list.length > 0); }
 
 /**Gets the code declaration thats been parsed through markdown-it.
  * @param info {TypeInfo | FieldInfo | PropertyInfo | EventInfo | MethodInfo} - The info to look into.